fix the path to the "config.h" file
Using the normal ./configure script, the "config.h" file would usually be
generated into the root dir and thus what was here before would have been
fine. But now, we put the headers into a separate config dir and include
that directory in the include path, so a relative include works.